move ext project component to its app
