+++ meinberlin/apps/dashboard/urls.py
@@ -2,6 +2,7 @@ from django.conf.urls import url
 
 from meinberlin.apps.dashboard2.urls import \
     urlpatterns as a4dashboard_urlpatterns
+from meinberlin.apps.extprojects.views import ExternalProjectCreateView
 
 from . import views
 
@@ -18,7 +19,7 @@ urlpatterns = [
     # Overwrite the ProjectUpdateView with meinBerlin urls
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/blueprints/'
         r'external-project/$',
-        views.ExternalProjectCreateView.as_view(),
+        ExternalProjectCreateView.as_view(),
         name='external-project-create'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/blueprints/'
         r'bplan/$',
