+++ meinberlin/apps/dashboard/views.py
@@ -7,8 +7,7 @@ from meinberlin.apps.bplan import models as bplan_models
 from meinberlin.apps.dashboard2 import mixins as a4dashboard_mixins
 from meinberlin.apps.dashboard2.components.forms.views import \
     ProjectComponentFormView
-from meinberlin.apps.dashboard2.views import ProjectCreateView
-from meinberlin.apps.extprojects import models as extproject_models
+from meinberlin.apps.extprojects.views import ExternalProjectCreateView
 from meinberlin.apps.newsletters.forms import NewsletterForm
 from meinberlin.apps.newsletters.views import NewsletterCreateView
 from meinberlin.apps.organisations.models import Organisation
@@ -51,33 +50,6 @@ class DashboardNewsletterCreateView(a4dashboard_mixins.DashboardBaseMixin,
         return reverse(
             'a4dashboard:newsletter-create',
             kwargs={'organisation_slug': self.organisation.slug})
-
-
-class ExternalProjectCreateView(ProjectCreateView):
-
-    model = extproject_models.ExternalProject
-    slug_url_kwarg = 'project_slug'
-    blueprint_key = 'external-project'
-    form_class = forms.ExternalProjectCreateForm
-    template_name = 'meinberlin_dashboard/external_project_create_form.html'
-
-
-class ExternalProjectUpdateView(ProjectComponentFormView):
-
-    model = extproject_models.ExternalProject
-
-    def get_project(self, *args, **kwargs):
-        project = super().get_project(*args, **kwargs)
-        return project.externalproject
-
-    def get_object(self, queryset=None):
-        return self.project
-
-    def validate_object_project(self):
-        return True
-
-    def validate_object_module(self):
-        return True
 
 
 class BplanProjectCreateView(ExternalProjectCreateView):
